Fixed Pokemon Trainer CSS
DO NOT CHANGE THE ECHO FIGHTERS OF THE SEPARATED POKEMON FOR POKEMON TRAINER. IF YOU DO, IT BREAKS THE ROTATION BETWEEN THE POKEMON TRAINER POKEMON WHEN SELECTING WHO TO START AS WITH POKEMON TRAINER. I REPEAT, DO NOT FUCK WITH THIS. IF YOU WANT BULBASAUR AND BLASTOISE TO BE ECHOES, MAKE A POKEMON TRAINER ECHO.

Capiche? Capiche.